consistent test fixtures naming, sequential picosol tests
