+++ docker-compose-local-validator.yml
@@ -4,7 +4,6 @@ services:
     command: [
         "solana-test-validator",
         "-r", # Reset flag
-        "-w", # Warp slot
         "--account-dir",
         "/test-fixtures",
         "--bpf-program",
