+++ test-fixtures/bpsol-manager-fee.json
