+++ test-fixtures/bpsol-stake-pool.json
