+++ test-fixtures/jupsol-stake-pool.json
