--- test-fixtures/pico-sol-stake-pool.json
@@ -1,14 +1,0 @@
-{
-  "pubkey": "8Dv3hNYcEWEaa4qVx9BTN1Wfvtha1z8cWDUXb7KVACVe",
-  "account": {
-    "lamports": 5143440,
-    "data": [
-      "AeVDKQDgR3b61MLyI/oLdrrL9Abi0e7jN26ks177tp6EDDCz/+HxU24fJ73s+KX4grGiXXlMMHZOVn3ifFtRKrKSQBHMTxJqGt9Z8+ORKewb19QVkqDpmyC3QqkDmDRxCv8t5Tm1+cZsprYaE3+egWRprk+31+48NcEavkLmlyQ1YRFiOkYy7RYkBC3+47A0IrgZz5ZrFiihNiA17wgE/absDDla18UNF/qu18wub6yLB67QePPJIe4KC0uEXwsh6kI5ZmJ62eWx/bMWl36EF6dM6UVvW7dxm4maOrezvmfQnQbd9uHXZaGT2cvhRs7reawctIXtX1s3kTqM9YV+/wCpnOdPr8hsAABrEmhQi2IAAPACAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAZAAAAAAAAAAAAAAAAAAAAAAAAGQAAAAAAAAAAAAAAAAAAADoAwAAAAAAAAAAAAAAAAAAAAAA6AMAAAAAAAABAAAAAAAAAAAA6AMAAAAAAAABAAAAAAAAAABr3AKMi2IAAAsOqdu6bAAA",
-      "base64"
-    ],
-    "owner": "SP12tWFxD9oJsVWNavTTBZvMbA6gkAmxtVgxdqvyvhY",
-    "executable": false,
-    "rentEpoch": 18446744073709551615,
-    "space": 611
-  }
-}
