--- test-fixtures/pico-validator-stake.json
@@ -1,14 +1,0 @@
-{
-  "pubkey": "ALkqRsYmCtWE1hx6VQPqtwnxwA8ymB5h8FzBNrF32SSN",
-  "account": {
-    "lamports": 119611217998556,
-    "data": [
-      "AgAAAIDVIgAAAAAALxsbDTisADc8y4cG2VUTzCo1qiLNEnkT4pFibQMyRfovGxsNOKwANzzLhwbZVRPMKjWqIs0SeRPikWJtAzJF+gAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAACv+jgfPFCYtQr61E7V2VULM+VKePfsbrkSRpIMVVF2jXB0GKMlsAABXAgAAAAAAAP//////////AAAAAAAA0D9cHQYoyWwAAAA=",
-      "base64"
-    ],
-    "owner": "Stake11111111111111111111111111111111111111",
-    "executable": false,
-    "rentEpoch": 18446744073709551615,
-    "space": 200
-  }
-}
