+++ test-fixtures/picosol-mint.json
@@ -3,7 +3,7 @@
   "account": {
     "lamports": 1461600,
     "data": [
-      "AQAAAC8bGw04rAA3PMuHBtlVE8wqNaoizRJ5E+KRYm0DMkX6axJoUItiAAAJAQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA==",
+      "AQAAAC8bGw04rAA3PMuHBtlVE8wqNaoizRJ5E+KRYm0DMkX6q5xpGJKqAAAJAQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA==",
       "base64"
     ],
     "owner": "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA",
