+++ test-fixtures/picosol-reserve-account.json
@@ -1,7 +1,7 @@
 {
   "pubkey": "2ArodFTZhNqVWJT92qEGDxigAvouSo1kfgfEcC3KEWUK",
   "account": {
-    "lamports": 3000613461708,
+    "lamports": 7707999,
     "data": [
       "AQAAAIDVIgAAAAAALxsbDTisADc8y4cG2VUTzCo1qiLNEnkT4pFibQMyRfovGxsNOKwANzzLhwbZVRPMKjWqIs0SeRPikWJtAzJF+gAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=",
       "base64"
