+++ test-fixtures/picosol-validator-list.json
@@ -3,7 +3,7 @@
   "account": {
     "lamports": 1461600,
     "data": [
-      "AgEAAAABAAAAULAR2MdsAAAAAAAAAAAAAPACAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAr/o4HzxQmLUK+tRO1dlVCzPlSnj37G65EkaSDFVRdow==",
+      "AgEAAAABAAAAEC5/kmG/AAAAAAAAAAAAABMDAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAr/o4HzxQmLUK+tRO1dlVCzPlSnj37G65EkaSDFVRdow==",
       "base64"
     ],
     "owner": "SP12tWFxD9oJsVWNavTTBZvMbA6gkAmxtVgxdqvyvhY",
