+++ test-fixtures/picosol-validator-stake.json
@@ -1,0 +1,14 @@
+{
+  "pubkey": "ALkqRsYmCtWE1hx6VQPqtwnxwA8ymB5h8FzBNrF32SSN",
+  "account": {
+    "lamports": 210442783116351,
+    "data": [
+      "AgAAAIDVIgAAAAAALxsbDTisADc8y4cG2VUTzCo1qiLNEnkT4pFibQMyRfovGxsNOKwANzzLhwbZVRPMKjWqIs0SeRPikWJtAzJF+gAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAACv+jgfPFCYtQr61E7V2VULM+VKePfsbrkSRpIMVVF2jkFhckmG/AABMAgAAAAAAAP//////////AAAAAAAA0D8pMnElAAAAAAAAAAA=",
+      "base64"
+    ],
+    "owner": "Stake11111111111111111111111111111111111111",
+    "executable": false,
+    "rentEpoch": 18446744073709551615,
+    "space": 200
+  }
+}
