+++ test-fixtures/referral-picosol-token.json
@@ -1,5 +1,5 @@
 {
-  "pubkey": "5YSa7x36xXZC3fRRbh3tEGAeE3penVphkAhpwDi6GCJM",
+  "pubkey": "5v1v8oxvCK1CuK8YcwfifWQb8QLi7tNFDXgJMm4TfTLA",
   "account": {
     "lamports": 2039280,
     "data": [
