+++ ts/tests/test/update-validator-list-balance.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, assert } from "vitest";
-import { readTestFixturesJsonFile } from "./utils";
+import { readTestFixturesAccPk, readTestFixturesJsonFile } from "./utils";
 import * as kit from "@solana/kit";
 import {
   deserStakePool,
@@ -10,13 +10,13 @@ import {
 
 describe("update-validator-list-balance", async () => {
   it("update-validator-list-balance-sim-mainnet", async () => {
-    const accountJson = readTestFixturesJsonFile("jupSolStakePool");
+    const accountJson = readTestFixturesJsonFile("jupsol-stake-pool");
     const accountData = Buffer.from(accountJson.account.data[0], "base64");
     const bytes = new Uint8Array(accountData);
     const stakePoolHandle = deserStakePool(bytes);
     const stakePool = getStakePool(stakePoolHandle);
 
-    const validatorListJson = readTestFixturesJsonFile("validatorList");
+    const validatorListJson = readTestFixturesJsonFile("validator-list");
     const validatorListData = Buffer.from(
       validatorListJson.account.data[0],
       "base64"
@@ -27,7 +27,7 @@ describe("update-validator-list-balance", async () => {
     let ix = updateValidatorListBalanceIxFromStakePool(
       {
         program: "SPMBzsVUuoHA4Jm6KunbsotaahvVikZs1JyTW6iJvbn",
-        stakePool: "8VpRhuxa7sUUepdY3kQiTmX9rS5vx4WgaXiAnXq4KCtr",
+        stakePool: readTestFixturesAccPk("jupsol-stake-pool"),
       },
       stakePoolHandle,
       validatorListHandle,
